put data methods in index class
